ETL_pipeline_boxofficemojo.com/requirements.txt
kafka_python
psycopg2-binary
apache-airflow
unidecode
boxoffice_api
pyspark
pandas
confluent_kafka
